Working version

+++ MiniJira.Repository/ILogRepository.cs
@@ -5,5 +5,5 @@ namespace MiniJira.Repository;
 public interface ILogRepository
 {
     Task CreateLog(long taskId, int userId, string description, CancellationToken cancellationToken);
-    Task<LogEntity[]> GetLogsByTaskId(int taskId, CancellationToken cancellationToken);
+    Task<LogEntity[]> GetLogsByTaskId(long taskId, CancellationToken cancellationToken);
 }
